adding features for traits among other 3rd job related features

+++ models/classInfo.go
@@ -2,7 +2,7 @@ package models
 
 type ClassInfo struct {
 	Name              string                `json:"name" firestore:"name"`
-	Level             int64                 `json:"level" firestore:"level"`
+	Level             int32                 `json:"level" firestore:"level"`
 	Exp               int64                 `json:"exp" firestore:"exp"`
 	CurrentWeapon     *string               `json:"currentWeapon,omitempty" firestore:"currentWeapon,omitempty"`
 	OldEquipmentSheet *OldEquipmentSystem   `json:"equipment,omitempty" firestore:"equipment,omitempty"`
